bunch of new and updated stuff

+++ Smart_Home/SmartHome/SmartHome/Source/HAL/Servo/Include/servo.h
@@ -14,6 +14,7 @@
 #include "dio.h"
 
 #define SERVO_TOP_VALUE_50_HZ			(4999)
+
 #define SERVO_ANGLE_0_DEGREE			(350)
 #define SERVO_ANGLE_90_DEGREE			(600)
 #define SERVO_ANGLE_180_DEGREE			(130)
@@ -21,8 +22,8 @@
 /*Function prototype*/
 
 /*Servo Initialization*/
-void servo_init(u8 portId, u8 pinId, u8 direction, u8 modeA, u8 modeB, u8 freq, u8 topValue);
+void servo_init(u8 portId, u8 pinId, u8 direction, u8 modeA, u8 modeB, u8 freq, u16 topValue);
 /*Set angle of servo*/
-void servo_setAngleDegree(u8 degree);
+void servo_setAngleDegree(u16 degree);
 
 #endif /* SERVO_H_ */
